Updated initalizer so it doesn't remake database over and over unless a boolean is set

+++ ENSE600_Project_1/src/main/java/DBInitializer.java
@@ -17,6 +17,8 @@ public class DBInitializer {
     private final Connection conn;
     private Statement statement;
     
+    private boolean REFRESH_MODE = false;
+    
     public DBInitializer() {
         dbManager = new DBManager();
         conn = dbManager.getConnection();
@@ -31,7 +33,7 @@ public class DBInitializer {
     }
     
     public void createItemsTable() {
-        if (dbManager.tableExists("Items")) {
+        if (dbManager.tableExists("Items") && REFRESH_MODE) {
             String sqlDropTable = "DROP TABLE Items";
             dbManager.updateDB(sqlDropTable);
         }
@@ -54,7 +56,7 @@ public class DBInitializer {
     }
     
     public void createEffectTypesTable() {
-        if (dbManager.tableExists("Effect_Types")) {
+        if (dbManager.tableExists("Effect_Types") && REFRESH_MODE) {
             String sqlDropTable = "DROP TABLE Effect_Types";
             dbManager.updateDB(sqlDropTable);
         }
@@ -72,7 +74,7 @@ public class DBInitializer {
     }
     
     public void createWeaponsTable() {
-        if (dbManager.tableExists("Weapons")) {
+        if (dbManager.tableExists("Weapons") && REFRESH_MODE) {
             String sqlDropTable = "DROP TABLE Weapons";
             dbManager.updateDB(sqlDropTable);
         }
@@ -92,7 +94,7 @@ public class DBInitializer {
     }
     
     public void createPotionsTable() {
-        if (dbManager.tableExists("Potions")) {
+        if (dbManager.tableExists("Potions") && REFRESH_MODE) {
             String sqlDropTable = "DROP TABLE Potions";
             dbManager.updateDB(sqlDropTable);
         }
